Move `Identity` to `:legacy:account` module
