+++ legacy/account/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id(ThunderbirdPlugins.Library.android)
+    alias(libs.plugins.kotlin.parcelize)
 }
 
 android {
