+++ legacy/core/src/test/java/com/fsck/k9/message/MessageBuilderTest.java
@@ -15,7 +15,7 @@ import java.util.Map;
 import app.k9mail.core.android.testing.RobolectricTest;
 import com.fsck.k9.Account.QuoteStyle;
 import com.fsck.k9.CoreResourceProvider;
-import com.fsck.k9.Identity;
+import app.k9mail.legacy.account.Identity;
 import com.fsck.k9.TestCoreResourceProvider;
 import com.fsck.k9.mail.Address;
 import com.fsck.k9.mail.BodyPart;
