+++ legacy/core/src/test/java/com/fsck/k9/message/ReplyActionStrategyTest.kt
@@ -1,12 +1,12 @@
 package com.fsck.k9.message
 
+import app.k9mail.legacy.account.Identity
 import assertk.assertThat
 import assertk.assertions.containsExactly
 import assertk.assertions.isEmpty
 import assertk.assertions.isEqualTo
 import assertk.assertions.isNull
 import com.fsck.k9.Account
-import com.fsck.k9.Identity
 import com.fsck.k9.helper.ReplyToParser
 import com.fsck.k9.mail.buildMessage
 import org.junit.Test
