+++ legacy/core/src/test/java/com/fsck/k9/notification/BaseNotificationDataCreatorTest.kt
@@ -1,12 +1,12 @@
 package com.fsck.k9.notification
 
+import app.k9mail.legacy.account.Identity
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isInstanceOf
 import assertk.assertions.isNotNull
 import assertk.assertions.isSameInstanceAs
 import com.fsck.k9.Account
-import com.fsck.k9.Identity
 import com.fsck.k9.K9
 import com.fsck.k9.K9.LockScreenNotificationVisibility
 import com.fsck.k9.NotificationLight
