+++ legacy/ui/legacy/src/main/java/com/fsck/k9/activity/AccountList.kt
@@ -9,8 +9,8 @@ import android.widget.ArrayAdapter
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import com.fsck.k9.Account
 import app.k9mail.legacy.account.BaseAccount
+import com.fsck.k9.Account
 import com.fsck.k9.K9.isShowUnifiedInbox
 import com.fsck.k9.Preferences.Companion.getPreferences
 import com.fsck.k9.search.SearchAccount.Companion.createUnifiedInboxAccount
