+++ legacy/ui/legacy/src/main/java/com/fsck/k9/activity/EditIdentity.kt
@@ -9,9 +9,9 @@ import androidx.core.content.IntentCompat
 import androidx.core.os.BundleCompat
 import androidx.core.view.isVisible
 import androidx.core.widget.doAfterTextChanged
+import app.k9mail.legacy.account.Identity
 import com.fsck.k9.Account
 import com.fsck.k9.EmailAddressValidator
-import com.fsck.k9.Identity
 import com.fsck.k9.Preferences
 import com.fsck.k9.ui.R
 import com.fsck.k9.ui.base.K9Activity
