+++ legacy/ui/legacy/src/main/java/com/fsck/k9/activity/compose/IdentityAdapter.java
@@ -13,7 +13,7 @@ import android.widget.BaseAdapter;
 
 import com.fsck.k9.Account;
 import com.fsck.k9.DI;
-import com.fsck.k9.Identity;
+import app.k9mail.legacy.account.Identity;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.ui.R;
 import com.fsck.k9.ui.identity.IdentityFormatter;
