+++ legacy/ui/legacy/src/main/java/com/fsck/k9/activity/compose/ReplyToPresenter.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.activity.compose
 
 import android.os.Bundle
-import com.fsck.k9.Identity
+import app.k9mail.legacy.account.Identity
 import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.Message
