+++ legacy/ui/legacy/src/test/java/com/fsck/k9/activity/compose/ReplyToPresenterTest.kt
@@ -2,11 +2,11 @@ package com.fsck.k9.activity.compose
 
 import android.os.Bundle
 import app.k9mail.core.android.testing.RobolectricTest
+import app.k9mail.legacy.account.Identity
 import assertk.assertThat
 import assertk.assertions.isFalse
 import assertk.assertions.isSameInstanceAs
 import assertk.assertions.isTrue
-import com.fsck.k9.Identity
 import com.fsck.k9.mail.Address
 import org.junit.Test
 import org.mockito.kotlin.doReturn
