+++ legacy/ui/legacy/src/test/java/com/fsck/k9/ui/messagedetails/MessageDetailsParticipantFormatterTest.kt
@@ -5,6 +5,7 @@ import android.text.Spannable
 import android.text.style.ForegroundColorSpan
 import androidx.core.text.getSpans
 import app.k9mail.core.android.testing.RobolectricTest
+import app.k9mail.legacy.account.Identity
 import assertk.assertThat
 import assertk.assertions.containsExactly
 import assertk.assertions.isEqualTo
@@ -12,7 +13,6 @@ import assertk.assertions.isInstanceOf
 import assertk.assertions.isNotNull
 import assertk.assertions.isNull
 import com.fsck.k9.Account
-import com.fsck.k9.Identity
 import com.fsck.k9.helper.ContactNameProvider
 import com.fsck.k9.mail.Address
 import org.junit.Test
