+++ legacy/ui/legacy/src/test/java/com/fsck/k9/ui/messageview/DisplayRecipientsExtractorTest.kt
@@ -1,9 +1,9 @@
 package com.fsck.k9.ui.messageview
 
+import app.k9mail.legacy.account.Identity
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import com.fsck.k9.Account
-import com.fsck.k9.Identity
 import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.buildMessage
 import org.junit.Test
