+++ legacy/ui/legacy/src/test/java/com/fsck/k9/ui/messageview/MessageViewRecipientFormatterTest.kt
@@ -5,12 +5,12 @@ import android.text.Spannable
 import android.text.style.ForegroundColorSpan
 import androidx.core.text.getSpans
 import app.k9mail.core.android.testing.RobolectricTest
+import app.k9mail.legacy.account.Identity
 import assertk.assertThat
 import assertk.assertions.containsExactly
 import assertk.assertions.isEqualTo
 import assertk.assertions.isInstanceOf
 import com.fsck.k9.Account
-import com.fsck.k9.Identity
 import com.fsck.k9.helper.ContactNameProvider
 import com.fsck.k9.mail.Address
 import org.junit.Test
